refactor: global vars
