+++ packages/fx-core/src/component/coordinator/index.ts
@@ -31,7 +31,7 @@ import { TelemetryEvent, TelemetryProperty } from "../../common/telemetry";
 import { MetadataV3 } from "../../common/versionMetadata";
 import { environmentNameManager } from "../../core/environmentName";
 import { ObjectIsUndefinedError } from "../../core/error";
-import { ErrorContextMW, globalVars } from "../../core/globalVars";
+import { ErrorContextMW, globalVars } from "../../common/globalVars";
 import { ResourceGroupConflictError, SelectSubscriptionError } from "../../error/azure";
 import {
   InputValidationError,
