+++ packages/fx-core/src/component/middleware/envMW.ts
@@ -5,7 +5,7 @@ import { Inputs, err } from "@microsoft/teamsfx-api";
 import _ from "lodash";
 import { environmentNameManager } from "../../core/environmentName";
 import { NoProjectOpenedError } from "../../core/error";
-import { TOOLS } from "../../core/globalVars";
+import { TOOLS } from "../../common/globalVars";
 import { CoreHookContext } from "../../core/types";
 import { QuestionNames } from "../../question/constants";
 import { selectTargetEnvQuestion } from "../../question/other";
