+++ packages/fx-core/src/component/utils/envUtil.ts
@@ -7,7 +7,7 @@ import { cloneDeep, merge } from "lodash";
 import { settingsUtil } from "./settingsUtil";
 import { LocalCrypto } from "../../core/crypto";
 import { pathUtils } from "./pathUtils";
-import { globalVars, TOOLS } from "../../core/globalVars";
+import { globalVars, TOOLS } from "../../common/globalVars";
 import * as path from "path";
 import { EOL } from "os";
 import { TelemetryEvent } from "../../common/telemetry";
