+++ packages/fx-core/src/core/middleware/projectMigratorV3.ts
@@ -16,7 +16,7 @@ import {
   sendTelemetryEvent,
   TelemetryEvent,
 } from "../../common/telemetry";
-import { TOOLS } from "../globalVars";
+import { TOOLS } from "../../common/globalVars";
 import {
   UpgradeV3CanceledError,
   MigrationError,
