+++ packages/fx-core/src/error/yml.ts
@@ -6,7 +6,7 @@
  */
 import { UserError, UserErrorOptions } from "@microsoft/teamsfx-api";
 import { getDefaultString, getLocalizedString } from "../common/localizeUtils";
-import { globalVars } from "../core/globalVars";
+import { globalVars } from "../common/globalVars";
 import { ErrorCategory } from "./types";
 
 /**
