+++ packages/fx-core/tests/component/developerPortalScaffoldUtils.test.ts
@@ -9,6 +9,7 @@ import "mocha";
 import path from "path";
 import * as sinon from "sinon";
 import { CapabilityOptions, getProjectTypeAndCapability } from "../../src";
+import { createContextV3, setTools } from "../../src/common/globalVars";
 import { developerPortalScaffoldUtils } from "../../src/component/developerPortalScaffoldUtils";
 import * as appStudio from "../../src/component/driver/teamsApp/appStudio";
 import {
@@ -24,10 +25,8 @@ import { MessagingExtension } from "../../src/component/driver/teamsApp/interfac
 import { StaticTab } from "../../src/component/driver/teamsApp/interfaces/appdefinitions/staticTab";
 import { manifestUtils } from "../../src/component/driver/teamsApp/utils/ManifestUtils";
 import { CommandScope, MeetingsContext } from "../../src/component/driver/teamsApp/utils/utils";
-import { createContextV3 } from "../../src/component/utils";
 import { DotenvOutput, envUtil } from "../../src/component/utils/envUtil";
 import { ObjectIsUndefinedError } from "../../src/core/error";
-import { setTools } from "../../src/core/globalVars";
 import { QuestionNames } from "../../src/question/constants";
 import { MockTools } from "../core/utils";
 import { MockedAzureAccountProvider, MockedM365Provider } from "../plugins/solution/util";
