+++ packages/fx-core/tests/component/driver/arm/bicepChecker.test.ts
@@ -1,18 +1,16 @@
-import "mocha";
+import { Context } from "@microsoft/teamsfx-api";
+import { AxiosRequestConfig, default as axios } from "axios";
 import * as chai from "chai";
 import chaiAsPromised from "chai-as-promised";
-import * as sinon from "sinon";
+import "mocha";
 import mockFs from "mock-fs";
-import { AxiosRequestConfig, default as axios } from "axios";
+import * as sinon from "sinon";
 import * as stream from "stream";
-
+import { createContextV3, setTools } from "../../../../src/common/globalVars";
 import { ensureBicepForDriver } from "../../../../src/component/driver/arm/util/bicepChecker";
-import { cpUtils } from "../../../../src/component/utils/depsChecker/cpUtils";
-import { createContextV3 } from "../../../../src/component/utils";
-import { MockTools } from "../../../core/utils";
-import { setTools } from "../../../../src/core/globalVars";
-import { Context } from "@microsoft/teamsfx-api";
 import { DriverContext } from "../../../../src/component/driver/interface/commonArgs";
+import { cpUtils } from "../../../../src/component/utils/depsChecker/cpUtils";
+import { MockTools } from "../../../core/utils";
 
 chai.use(chaiAsPromised);
