+++ packages/fx-core/tests/component/generator/officeXMLAddinGenerator.test.ts
@@ -16,6 +16,7 @@ import { OfficeAddinManifest } from "office-addin-manifest";
 import * as path from "path";
 import * as sinon from "sinon";
 import * as uuid from "uuid";
+import { createContextV3, setTools } from "../../../src/common/globalVars";
 import { cpUtils } from "../../../src/component/deps-checker/";
 import { Generator } from "../../../src/component/generator/generator";
 import { HelperMethods } from "../../../src/component/generator/officeAddin/helperMethods";
@@ -23,8 +24,6 @@ import {
   OfficeXMLAddinGenerator,
   OfficeXmlAddinGeneratorNew,
 } from "../../../src/component/generator/officeXMLAddin/generator";
-import { createContextV3 } from "../../../src/component/utils";
-import { setTools } from "../../../src/core/globalVars";
 import {
   OfficeAddinHostOptions,
   ProgrammingLanguage,
