+++ packages/fx-core/tests/plugins/resource/spfx/depsChecker/generatorChecker.test.ts
@@ -1,16 +1,15 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import { Colors, LogLevel, LogProvider, UserError } from "@microsoft/teamsfx-api";
+import { LogLevel, LogProvider, UserError } from "@microsoft/teamsfx-api";
 import chai from "chai";
 import fs from "fs-extra";
 import "mocha";
 import { restore, stub } from "sinon";
+import { createContextV3, setTools } from "../../../../../src/common/globalVars";
 import { cpUtils } from "../../../../../src/component/deps-checker/util/cpUtils";
 import { GeneratorChecker } from "../../../../../src/component/generator/spfx/depsChecker/generatorChecker";
 import { telemetryHelper } from "../../../../../src/component/generator/spfx/utils/telemetry-helper";
-import { createContextV3 } from "../../../../../src/component/utils";
-import { setTools } from "../../../../../src/core/globalVars";
 import { MockTools } from "../../../../core/utils";
 
 class StubLogger implements LogProvider {
