+++ packages/fx-core/tests/question/create.test.ts
@@ -25,6 +25,8 @@ import * as path from "path";
 import sinon from "sinon";
 import { FeatureFlagName } from "../../src/common/constants";
 import { isApiCopilotPluginEnabled } from "../../src/common/featureFlags";
+import * as utils from "../../src/common/globalVars";
+import { setTools } from "../../src/common/globalVars";
 import { getLocalizedString } from "../../src/common/localizeUtils";
 import { sampleProvider } from "../../src/common/samples";
 import { AppDefinition } from "../../src/component/driver/teamsApp/interfaces/appdefinitions/appDefinition";
@@ -32,8 +34,6 @@ import { manifestUtils } from "../../src/component/driver/teamsApp/utils/Manifes
 import { pluginManifestUtils } from "../../src/component/driver/teamsApp/utils/PluginManifestUtils";
 import { OfficeAddinProjectConfig } from "../../src/component/generator/officeXMLAddin/projectConfig";
 import { convertToLangKey } from "../../src/component/generator/utils";
-import * as utils from "../../src/component/utils";
-import { setTools } from "../../src/core/globalVars";
 import { FileNotFoundError } from "../../src/error";
 import {
   ApiMessageExtensionAuthOptions,
